TDP-1559: Prevent choosing integer over double during full analysis.

+++ dataprep-backend-common/src/main/java/org/talend/dataprep/api/type/Type.java
@@ -33,7 +33,7 @@ public enum Type implements Serializable {
     DATE("date", STRING, "DATE"), //$NON-NLS-1$
     NUMERIC("numeric", STRING, "NUMERIC"), //$NON-NLS-1$
     DOUBLE("double", NUMERIC, "DOUBLE"), //$NON-NLS-1$
-    INTEGER("integer", DOUBLE, "INTEGER"), //$NON-NLS-1$
+    INTEGER("integer", NUMERIC, "INTEGER"), //$NON-NLS-1$
     FLOAT("float", NUMERIC, "FLOAT"); //$NON-NLS-1$
 
     /** Serialization UID. */
